Ms2 test

+++ app/src/main/java/com/example/birdsofafeather/FakedMessageListener.java
@@ -1,16 +1,12 @@
 package com.example.birdsofafeather;
 
-import android.content.Intent;
-import android.content.SharedPreferences;
 
 import com.example.birdsofafeather.model.db.AppDatabase;
 import com.example.birdsofafeather.model.db.Course;
 import com.example.birdsofafeather.model.db.Person;
-import com.example.birdsofafeather.model.db.PersonWithCourses;
 import com.example.birdsofafeather.model.db.Session;
 import com.google.android.gms.nearby.messages.Message;
 import com.google.android.gms.nearby.messages.MessageListener;
-import com.example.birdsofafeather.PersonListActivity;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -25,7 +21,6 @@ public class FakedMessageListener extends MessageListener {
     private final MessageListener messageListener;
     //instantiate executor to execute messages
     private final ScheduledExecutorService executor;
-    private Intent PersonListActivity;
 
     public FakedMessageListener(MessageListener realMessageListener, String messageStr, AppDatabase db, String sessionID) {
         this.messageListener = realMessageListener;
@@ -75,7 +70,7 @@ public class FakedMessageListener extends MessageListener {
                 //start use scanner to scan each line
                 while (scanner.hasNextLine()) {
                     String[] array = scanner.nextLine().split(csvSplitBy);
-                    String firstValue = array[0];
+                    //set UUID
                     if(count == 0) {
                         personId = array[0];
                         count++;
@@ -90,11 +85,14 @@ public class FakedMessageListener extends MessageListener {
                         count++;
                     //set profile courses
 
-                    }                    //TODO: Here should put the real ourself UUID
-                    else if (array[0].equals(selfId) || array[0].equals("4b295157-ba31-4f9f-8401-5d85d9cf659a")) {
-                        //TODO: set the waving to ourself to true
-                        isWave = true;
                     }
+                    else if (array[0].equals(selfId) || array[0].equals("4b295157-ba31-4f9f-8401-5d85d9cf659a")) {
+                        isWave = true;
+                        if(db.personWithCoursesDao().exists(personId)) {
+                            db.personWithCoursesDao().updateWavingToUs(true, personId);
+                        }
+                    }
+                    //create this person's courses
                     else if (array.length == 5){
                         year = array[0];
                         quarter = array[1];
@@ -105,9 +103,7 @@ public class FakedMessageListener extends MessageListener {
                         text = quarter + year + ' ' + courseType + ' ' + courseNum;
                         int courseId = db.coursesDao().maxId() + 1;
 
-                        //String personString = name + photoId;
-                        //personId = UUID.nameUUIDFromBytes(personString.getBytes()).toString();
-
+                        //Create the course, but only insert it into the db if the user also has this course
                         Course c = new Course(courseId, personId, text, year,quarter,courseSize);
                         if(userCourseText.contains(c.text) && !db.personWithCoursesDao().exists(personId)){
                             db.coursesDao().insert(c);
@@ -115,7 +111,7 @@ public class FakedMessageListener extends MessageListener {
                     }
                 }
 
-                boolean s = db.personWithCoursesDao().exists(personId);
+                //If the person doesn't already exist in the database, add them to it
                 if(!db.personWithCoursesDao().exists(personId)){
                     //set the student profile id
                     Person newPerson = new Person(personId, name, photoId, false);
@@ -181,22 +177,22 @@ public class FakedMessageListener extends MessageListener {
                         newPerson.sizePriority = sizePrio;
                         newPerson.recentPriority = recentPrio;
                         db.personWithCoursesDao().insert(newPerson);
-
-                        List<String> peopleInSession = db.sessionsDao().get(sessionID).peopleIDs;
-                        if(!peopleInSession.contains(personId)) {
-                            peopleInSession.add(personId);
-                            Session updatedSession = new Session(sessionID, db.sessionsDao().get(sessionID).sessionName);
-                            updatedSession.peopleIDs = peopleInSession;
-                            db.sessionsDao().delete(db.sessionsDao().get(sessionID));
-                            db.sessionsDao().insert(updatedSession);
-                        }
                     }
+                }
+
+                //add person into the current session
+                List<String> peopleInSession = db.sessionsDao().get(sessionID).peopleIDs;
+                if(db.personWithCoursesDao().exists(personId) && !peopleInSession.contains(personId)) {
+                    peopleInSession.add(personId);
+                    Session updatedSession = new Session(sessionID, db.sessionsDao().get(sessionID).sessionName);
+                    updatedSession.peopleIDs = peopleInSession;
+                    db.sessionsDao().delete(db.sessionsDao().get(sessionID));
+                    db.sessionsDao().insert(updatedSession);
                 }
 
                 scanner.close();
                 //send back the message that the profile is set up
                 this.messageListener.onLost(message);
-
             }
         });
     }
